feat: Add email sending to user creation process
This commit adds functionality to send an email notification when a new user is created. It imports the `Queue` class from the `bull/lib/queue` module and creates a new instance of the queue named "email sending". The `postNew` method in the `UsersController` class now checks if the email and password are provided in the request body. If either of them is missing, it returns a 400 error response. It also checks if the user already exists in the database and returns a 400 error response if they do. If the user is new, it inserts the user into the database with a hashed password and adds a job to the "email sending" queue. Finally, it returns a 201 response with the email and user ID.