refactor: Update UsersController to retrieve user by ObjectID
This commit updates the `UsersController` module to retrieve a user by their `ObjectID` instead of using the `ObjectId` constructor from the `dbClient`. It also imports the `ObjectID` class from the `mongodb` package. This change ensures consistent usage of the `ObjectID` and improves the efficiency of retrieving user data from the database.